* [FIX] Solves #450. Add some additional checks for missing items.

+++ inc/SP/Core/Upgrade/User.class.php
@@ -89,13 +89,13 @@ class User
             DB::getQuery($Data);
 
             $query = /** @lang SQL */
-                'DELETE FROM usrToGroups WHERE usertogroup_userId NOT IN (' . $paramsIn . ') OR usertogroup_userId IS NULL';
+                'DELETE FROM usrToGroups WHERE usertogroup_userId <> ? AND usertogroup_userId NOT IN (' . $paramsIn . ') OR usertogroup_userId IS NULL';
             $Data->setQuery($query);
 
             DB::getQuery($Data);
 
             $query = /** @lang SQL */
-                'DELETE FROM accUsers WHERE accuser_userId NOT IN (' . $paramsIn . ') OR accuser_userId IS NULL';
+                'DELETE FROM accUsers WHERE accuser_userId <> ? AND accuser_userId NOT IN (' . $paramsIn . ') OR accuser_userId IS NULL';
             $Data->setQuery($query);
 
             DB::getQuery($Data);
